Refactors the code into several modules
