--- src/msg.rs
@@ -1,4 +1,0 @@
-rosrust::rosmsg_include!(
-    tf2_msgs/TFMessage, 
-    geometry_msgs/TransformStamped, 
-    geometry_msgs/Transform);
